Исправила замечания
